# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Operator360 Portal Backend API
  version: 1.0.0
  description: |
    Contract reference for the Operator360 portal backend, generated from the
    current handler implementations on branch f/oid-pid. All routes are under
    `/api` and require a bearer/session token validated by AuthMiddleware.

    Endpoints whose S3 file paths are keyed by an operator now resolve the S3
    prefix via `opt_master.data_path` looked up by `opt_id` — they no longer
    accept or require `opt_state` / `opt_district`.
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required: [error]
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_records:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_previous:
          type: boolean
...
  responses:
    Unauthorized:
      description: User not found in context
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    ServerError:
      description: Internal server error (S3, DB, or parse failure)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  parameters:
    OptId:
      name: opt_id
      in: query
      required: true
      schema:
        type: string
      example: MH_WMIT_SN_NS046496
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        maximum: 1000
paths:
...
  /operator_feedback_stats:
    get:
      summary: Get an operator's feedback submission history and per-category stats
      description: |
        Reads every /feedback (see the /feedback POST endpoint) submission for
        this opt_id back out of ClickHouse, where a Kafka Engine + materialized
        view flattens each feedback_submitted event into one row. Returns
        yes/no/unset counts per verdict category plus the full submission
        history, for the operator profile page's feedback stats card.
      operationId: GetOperatorFeedbackStats
      parameters:
        - $ref: "#/components/parameters/OptId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  opt_id: { type: string }
                  total_feedback_count: { type: integer }
                  categories:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        yes: { type: integer }
                        no: { type: integer }
                        unset: { type: integer }
                  latest: { type: object, nullable: true, additionalProperties: true }
                  history:
                    type: array
                    items: { type: object, additionalProperties: true }
        "400":
          description: opt_id missing
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /anamolous_sids:
    get:
      summary: List/paginate an operator's anomalous SIDs
      description: |
        File path is resolved via opt_master.data_path looked up by opt_id.
        opt_state/opt_district are no longer accepted.
      operationId: GetAnamolousSIDs
      parameters:
        - $ref: "#/components/parameters/OptId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: anomaly_category
          in: query
          schema: { type: string }
          example: Biometrics
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  regional_office: { type: string }
                  operator_id: { type: string }
                  file: { type: string }
                  anomaly_category_filter: { type: string }
                  pagination: { $ref: "#/components/schemas/Pagination" }
                  count: { type: integer }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        sid: { type: string }
                        operator_id: { type: string }
                        anomaly_category: { type: string }
                      additionalProperties: true
                  requested_by: { type: string }
        "400":
          description: opt_id missing
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Operator data path or anomaly_sid.json not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "500": { $ref: "#/components/responses/ServerError" }
...